Implement better classification of character
A character should be classified as being in the wrong spot only if that
character still remains in the word. For example, the word to guess is
"beast" and the player guesses "beret". The first 'e' is in the correct
spot but the second 'e' should not be classified as in the wrong spot
because there is no more 'e' in the word to guess.